+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 4.13
 
 - [#1830](https://github.com/openshift/cluster-monitoring-operator/pull/1830) Add alert KubePodNotScheduled
-- [#1843](https://github.com/openshift/cluster-monitoring-operator/pull/1830) Node Exporter ignores network interface under name "enP.*".
+- [#1843](https://github.com/openshift/cluster-monitoring-operator/pull/1843) Node Exporter ignores network interface under name "enP.*".
+- [#1860](https://github.com/openshift/cluster-monitoring-operator/pull/1860) Adds runbook for PrometheusRuleFailures
 
 
 ## 4.12